make the exporter extendable
